changed double variable without capital

+++ src/unit1project.java
@@ -2,14 +2,14 @@ import java.util.Scanner;
 public class unit1project {
     public static void main(String[] args){
      Scanner s = new Scanner(System.in);
-     Double noTaxBill;
-     Double tipPercentage;
-     Double tipAmount;
-     Double totalTipAmount;
+     double noTaxBill;
+     double tipPercentage;
+     double tipAmount;
+     double totalTipAmount;
      int partyAmount;
-     Double totalBill;
-     Double personTip;
-     Double  eachPersonAmount;
+     double totalBill;
+     double personTip;
+     double  eachPersonAmount;
      System.out.println("");
         System.out.println("Hello User! Please input the bill amount, tip percentage as a whole number, and the number of people in the party ");
         noTaxBill = s.nextDouble();
@@ -20,7 +20,7 @@ public class unit1project {
         totalBill = totalTipAmount + noTaxBill;
         personTip = totalTipAmount/partyAmount;
         eachPersonAmount = totalBill/partyAmount;
-        eachPersonAmount = Double.valueOf(Math.round(eachPersonAmount*100)/100);
+        eachPersonAmount = Math.round(eachPersonAmount*100)/100;
         System.out.print("Thank you. Your total tip amount is " + totalTipAmount + " Your total bill with tip is: " + totalBill + " Each person must pay " + personTip+ " each for tip. The total per person is: " + eachPersonAmount);
